Inserido regra de calculo de salario, alteração para visualização do formulario de retorno

+++ TaxingArchitecture/Tax.ViewModels/EmployeeViewModel.cs
@@ -4,8 +4,8 @@
     {
         public string Nome { get; set; }
         public string Email { get; set; }
-        public double SalarioBruto { get; set; }
-        public double SalarioLiquido { get; set; }
-        public double Imposto { get; set; }
+        public decimal SalarioBruto { get; set; }
+        public decimal SalarioLiquido { get; set; }
+        public decimal Imposto { get; set; }
     }
 }
